requirements.txt
aiogram
python-dotenv
aiohttp
requests
gunicorn
flask